% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Penguin measurements dataset}
\format{
A data frame with the following columns:
\describe{
\item{species}{Species of penguin (Adelie, Chinstrap, Gentoo)}
\item{island}{Island where the penguin was observed}
\item{sex}{Male or Female}
\item{year}{Observation year}
\item{flipper_length_mm}{Flipper length (millimetres)}
\item{body_mass_g}{Body mass (grams)}
\item{bill_length_mm}{Bill length (millimetres)}
\item{bill_depth_mm}{Bill depth (millimetres)}
}
}
\source{
Adapted from the palmerpenguins dataset
}
\usage{
penguins
}
\description{
This dataset contains body measurements for penguins observed in the Palmer Archipelago.
}
\keyword{datasets}
