% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launched the Shiny app}
\usage{
run_app()
}
\value{
Invisibly returns the result of shiny::runApp().
}
\description{
Opened the interactive Penguin Explorer included in this package.
}
